Add-on update check: check all add-ons at once in order to simulate responsive REST API.
Update checks will be done for all add-ons at once instead of each add-on seek out its updated versions. This is done by letting checkForAddonUpdate function take in a 'dictionary' that'll be looped through, and for each entry, update checking will be done. This simulates REST API by letting the client pass in the entire content at once instead of using separate connection attempts for map entries. In NVDA Core version, this simulation function will be replaced by sending and receiving add-on update data between the client and the server.