Support automatic add-on updating.
In addition to manual add-on updates, NVDA can be told to check for add-on updates automatically. This is controlled by the new 'Add-on Updater' panel in NVDA settings.
When new add-on updates are available via automatic update checking, present a message informing the user the add-on updates are available. This is a necessary step so add-on updates list can show up properly.