Update checks: allow users to specify some add-ons should prefer development updates.
Parlty to support the next add-on to be added for update checks and to answer a question from several add-on authors: add a checkable list to let users specify that some add-ons should receive updates from development channel.
Also take care of a regexp problem where add-on name wasn't part of the resulting version string.