Update check: simplify add-on URL's.
Implementation detail: add-ons hosted on community add-ons site are idenfied with a unique alphanumeric key. Thankfully, the URL to access these is the same (the difference is the key). Hence simplify URL listing by just listing add-on keys instead of the full URL, which will be constructed while updatecheck threads are running.