Update fetcher: split addresses via slash character.
All the info we need to locate add-on version info can be found in the tail end of the URL - after the last slash regardless of whether this is an add-on from community add-ons server or not. Also, instead of using resource.url, just assign a local add-on URL variable.