Update check: fetch the currently hosted add-ons map via JSON, also resolving five minute timeout issue for GitHub URL's.
Previously, a number of threads were used to access individual keys for the installed add-on found on add-on server. Although this method did skip add-on keys that are missing from add-ons server, it would return temporary URL's for some add-on downloads, particularly for ones hosted on GitHub. As a result, download count was inflated, and led to timeout issues (fivem inute timeout for GitHub links).
Therefore use json endpoint provided by the add-on server which can be used to transfer add-ons list from the server to the client. Not only this keeps download count in check, it will not result in GitHub being accessed, and therefore leading to fetching the URL as listed on the server, not the temporary GitHub URL's. This also resolves timeout problem - one possible bug might be a timing issue between fetch and download, but this is taken care of via error messages.
As always, because the fetching routine uses urllib, it must run in a separate thread (an internal function will be used) and joined, because without joining, results will be lost.