Legacy add-on detection announcement: present an announcement if legacy add-ons are found, and if so, disable automatic add-on update check for this session.
At startup, present a message if legacy add-ons are found, and if so, disable automatic add-on update check.
Specifically, after detecting legacy add-ons, reset legacy add-ons set based on a set intersection:
* Current set of legacy add-ons found.
* A set of legacy add-ons detected at the moment.
Intersection is useful for clearing out legacy add-ons set NVDA is holding onto (via update state), especially if a legacy add-on was removed. This mechanism also allows legacy add-on message to be announced when new legacy add-ons are detected, not old ones already detected.