Update fetcher: traverse another URL if initial add-on update URL does not contain '.nvda-addon'.
Previously another URL was fetched if dev builds of certain add-ons (notably StationPlaylist and Windows 10 App Essentials) are sought because the URL does not incolude '.nvda-addon' at the end. Therefore generalize this to any add-on result URL. Although this might hurt performance (as no update channel string check will be performed), it allows the add-on to respond to similar cases later if needed.