Add-on update check: raise an error if IOError other than SSL error occurs.
Make sure to handle IOError stemming from:
* Wrong add-on update URL.
* No Internet connection.
If an error does occur, record this fact in results dictionary so the main update check function can raise a runtime error and the exception logged. This will result in an error dialog shown to users if manual add-on update check is in progress, similar to NVDA Core update check.