Toast notifications: notify add-on update availability via Windows 10 toasts.
Call wx.adv.NotificationMessage.Show to display toasts when add-on updates are available. Due to technical limitations, this will be used on Windows 10 only, and when toasts appear, add-on updates dialog will be shown directly when a user visits nVDA menu/Tools/Check for add-on updates (done by storing update info inside a dedicated private dictionary).