Legacy add-ons dilaog: use a dedicated dialog to present legacy add-ons and reasons.
Inspired by incompatible add-ons dilaog found in NVDA: use a dedicated dialog to display legacy add-ons and reasons for declaring them as such. The ideal place for this dialog is extended add-on GUI module, but it is placed in the top-level global plugin (__init__) entry point in case legacy add-ons are found while running NVDA from source code (add-on updates are turned off in this case).