Update check: add a function to retrieve community add-ons metadata JSON file.
A JSON file will be used to record add-on metadata such as add-on summary, author, minimum and lasted tested NVDA versions, and add-on update key found in add-on files's get.php file. For ease of maintenance, this JSON file will live in nvdaaddons.github.io. Because JSON retrieval uses urllib, create a dedicated function to retrieve the JSON file that will be run inside another thread.