Update check: fetch add-on update keys for new add-ons via add-ons metadata.
With more add-ons coming to community add-ons website and added to get.php file, it becomes harder to track which add-on identifier is assigned to add-on update key from get.php file. Fortunately, the new community add-ons metadata includes addonKey field used to teach Add-on Updater about identifier:update key mappings. Therefore use this if metadata can be used, and if not, check classic add-on names to URL's map, and failing that, give up.