Add-on Updater settings: add add-on update notification combo box.
User suggestion: can NVDA present a dialog instead of a toast when add-on updates are available the first time nVDA runs? In a way, yes, although for simplicity this setting will be in effect throughout an nVDA session.
A new setting named 'add-on update notification' was added to let Windows 10 and later users see a toast or a dialog when add-on updates are available (values themselves are stored as strings). By default, notification toast will be displayed.