Update check: use SSLCertVerificationError instead of SSLError class.
On September 30, 2021, users began reporting that NVDA and add-on update checks were failing with an error message, which was traced to certificate verification failure (possible effect of expired Let's Encrypt root certifcate). It turns out this issue was resolved in NVDA 2021.3 (alpha at the moment) with an August 2021 alpha build, therefore 'backport' (apply) the fix (using SSLCertVerificationError instead of SSLError) in Add-on Updater as it is also affected.