Readme: add a note on inability to complete add-on update check functionality on domain-joined systems.
Reported by a user: on Windows 7, Add-on Updater does not work properly and says updates are not available when there are add-on updates. An add-on author pointed out that this might be case if the computer is joined to a corporate domain. Therefore add a note about this in the readme.