Add-on update proc/update check: return update records list instead of a dictionary of update metadata.
Project Meteor: return a list of update records instead of an update metadata map in order to take advantage of add-on update record objects and obtain needed properties. Becaseu update record class includes a method to return a dictionary view of update metadata, it is possible to return either the object directly or a metadata dictionary.