Add-on update check: define a new extension point action for changing update check menu item.
It does not make sense to call Add-on Updater menu item 'check for add-on updates' when updates are available (in some ways, keeping it as is makes sense for some as it is after all checking for updates and show results). Therefore introduce a mechanism to change update menu item through an extension point action that simply changes menu item label (thankfully menu item is stored as a global plugin attribute so changing labels is possible). This extension point action will take menu item label and change it.