Add-on handler: check for add-on updates from add-on update proc module.
Use ad-on update proc version of checkForAddonUpdates fnction to check for add-on updates. Unlike add-on handler version, the new update proc version returns a list of add-on update records which should be converted into a dictioanry for compatibility with add-on updates dialog. The conversion routine (next commit) will be removed once add-on updates dilaog is edited to work directly with update records list.