Add-on installer: use install add-on function provided by ad-on update proc with changes in progress bar dialog presentation.
Add-on update proc includes install add-on function which duplicates internals of add-on instal routine such as opening bundles, checking compatibility flag, and raising errors if problems occur. As the new routine is an internal function, it does not present dialogs but returns install status enumeration. Therefore edit the dedicated add-on install method from downloader class as follows:
1. Progress dialog will be shown throughout the update process even if compatibility erros ocur.
2. Install status enumeration values will be checked.
3. Add-on bundle routines are gone as the ad-on update proc's instlal add-on function deals with the add-on bundle directly.
4. Removed try statements as the downloader class method does not have to deal with bundle errors and to make the code easier to read.
5. Instead of bundle name, add-on name as seen by downloader class wil be appended to updated add-ons list.