Add-on downloader/install: completely separate add-on install rotine from downloader class.
One integration goal of Project Meteor: add-on install routine is now a separate function from add-on downloader class. Note that update add-on routine is called immediately as temp path is gone when downoader class is gone and the next generator is called. This introduces artifacts from previous add-on update attempts if updating multiple add-ons.