Add-on download: define a minimal add-on downoad dialog separate from downloader class.
Another idea behidn Project Meteor: separate add-on download routine from the user interface. As part of this, add-on update proc contains download add-ons function which is a direct relaement (or close to a direct replacement) of download routine fond in add-on downloader class. In order to use the new routine (which doesn't update progress dialogs automatically), a separate function is required that houses a progress dialog and wraps arond add-on update proc's download function. A minimal function which also updates a 'downloaded ad-on bundle' given temp path and summary (name) is hereby defined to demonstrate this possibility.