Add-on download: minimal download progress action.
As add-on update proc's download function can notify listeners about download progress, have a dedicated listener to act when download size changes. Note that this blocks the main thread (NVDA may appear to not respond when keyboard keys are pressed as a result). At least this provides one result of separating add-on update download internals from the accompanying user interface.