Add-on download: define an update add-ons function to coordinate add-on downloads and installs.
Define a dedicated update add-ons function which takes a list of add-ons and auto boolean argument, similar to update add-ons generator. In fact, the new function is a direct replacement for the generator and is called from add-on updates dialog. Nopte that main thread is still blocked due to use of iteration to download and install add-ons one at a time.