Add-on GUI: define two functions to download and instal add-ons all at once.
Instead of downloading and installing one add-on at a time, define two functions to download and install all add-ons at once (download add-ons, followed by install all of them)). Note that for install loop, function name is installAddons since updateAddons function is used to coordinate add-on download and installs. With this change, a single download progress dialog will be used to inform download progress but this does result in blocking the main thread (no keyboard commands).