Add-on GUI: use a dedicated thread for downloading and installing add-on updates.
Because download progress dialog blocks the main thread, use a separate thread to download and install add-on updates. As part of this change, download progress dialog definition was transferred to update add-ons function with a separate thread called to call download updates function. In addition, add-on installation loop is called from the main thread via wx.CallAfter.