Add-on install loop: handle errors in a switch/case fashion while keeping track of update success count.
Handle install status flags in a switch/case (if/elif/else) fashion for readability (some log records are handled by add-on update proc's install add-on function anyway). Also, instead of using a list of updated add-ons, increment a successful install count when add-ons are indeed updated and use this to present restart dialog if required.