Add-on andler: remove unused names to urls dictionary.
As names to urls dictionary is part of add-on update proc module, there is no longer any need to keep a copy in extended add-on handler. In short, this removes a feature introduced in early days of Add-on Updater, specifically a mapping between add-on identifier and add-on update key (at least the mapping is found in a new module).