Add-on update check protocols: new add-on update protocols module.
An update check protocol is a class that retrieves add-on update metadata from specific sources using different formats. For example, a protocol may need to deal with a JSON dictionary or access multiple URL's to accomplish its task. Several common methods will be defined, the most important being a method to return update metadata in a common format. In short, update check protocol is a refactor update check routine collected under a single class. A module to host protocols (initially based on ad-on update proc module) is hereby defined to isolate it from the rest of the add-on.