Update check protocol: define protocol 0 (base/no updates).
A base protocol (0/base/no updates) is defined to serve as a blueprint for other protocols. This protocol can actually be instantiated but subclasses should be used. This protocol defines a check for add-on updates method which takes in a dicionary of instlaled addons metadata and returns a list of add-on update records defined in update proc module.