Update protocol 3: define a custom getAddonsData mthod to acocnt for JSON format differneces.
Unlike other protocols, Spanish community catalog returns a list formatted in JSON. This means the base implementation of getAddonsData iwll present an attribute error because update method is a dictionary method, not that of a list. Therefore define a custom getAddonsData method that will simulate dictionray creation by storing JSON data (list of add-on ID's and their metadata) inside a dedicated key.