Update protocol 3: eit fetch add-on info to work directly with Spanish community catalog.
All the info required to detect updates is housed inside Spanish community catalog JSON list. Therefore there is really no need to access other websites with an exception of community add-ons website for URL's not ending in '.nvda-addon' (same issue as protocol 2). The key change is that there is no need to subject URL's to pattern matching to determine update version as this info is recorded in the catalog (sometimes the version from the catalog and the one reported by the actual URL is different but this is more towards an issue with the catalog). At the moment there is no compatibility check but it will be done soon (future commit) as compatibility info is stored as strings, not integer tuple.