Update check proc: remove duplicate code from protocol 2.
As bulk of update procedures/processes module is part of protocol 2 now, there is no need to keep the duplicate in this module. Also clarify the role of update poc module - update proc is higher in the layer than update protocols as the procedures module calls upon protocols to perform update check task.