Update protocols: work directly with add-on update record attributes when fetching update information.
As fetch update info method takes add-on update records, be sure to use attributes such as version and url when checking for updates. The actual return value from this function is a list of update records with update availability (next). Also, unconditionally set update version (previously version string was set if installed and update versions are different). In additoin, since the fetch info takes in only the results dictionary and update record, there is no need to specify add-on name and manifest info anymore, significantly simplifying the fetch method.