Add-on handler/update: introduce ability to download and install updates in the background.
Simplified version of download and install user interface: update add-ons in the background from a separate thread. Due to impleentation issues (related to notifications), this feature is exclusive to Windows 10 and later client systems. With background updates, two loops will be used: one to download all add-ons, followed by install loop. The GUI will also be modified, mostly dealing with the control used to display add-ons in add-on updates dialog.