Update protocol 2/fetch update info: remove one last remaining link to names to URL's from protocol 2.
As add-ons found in names to URL's dictionary are now included in add-ons metadata, there is no need to check for membership in this dictionary (dead code). Besides, the code would not work anyway since the dictionary records strings as keys whereas the dead code asked for add-on record (incompatible data type). With this change, protocol 2 is close to independence from protocol 1 (community add-ons info without compatibility data) - the only link is results map recording add-on keys to download inks.