Add-on update protocols: introduce protocol 4 (NV Access add-ons datastore).
In 2020', NV Access designed an add-ons datastore that includes records of al add-ons submitted and their metadata such as version, compatibility information, and update channel and links. This is a significant improvement over protocols 1 (community add-ons website without compat info) nad 2 (website with compat info present). Due to similarities, the new protocol for the datastore (protocol 4) resembles Spanish community add-ons catalog (for both protocols, no need to fetch additional download URL information for most add-ons, and in case of protocol 4, no need to access additional sites at all). The key difference from other protocols (especially different from protocol 3) is that NV Access datastore returns ad-ons compatible with the NVDA version entered (as of time of this writing, the latest API is 2022.2 but no add-on papears to be updatable according to the datastore as it raises HTTP error 400 (bad request)). Note that this protocol is experimental, therefore it will be usable only through a command-line switch (future commit).