Add-on utilities: add clietn and Windows 10 check functions.
Preparing to end support for old Windows releases nad code cleanup: add two check functions to retrn if this is a Windows client (not a server computer) and/or is a Windows 10 or later system (not Windows Server 2016/2019/2022/other server releases). These functions are used to set automatic update flags to True or False depending on the result of these functions (for instance, automatic updates will be disabled on server systems) and for use from other components when performing update checks and settings presentation. This is a stepping stone toward removing bulk of the add-on's features from older Windows releases (7, 8, 8.1 and server variants) in 2023 (for example, automatic updates will be turned off on systems other than Windows 10 and 11 clients).