Add-on handler/update check: update records (info) must not be empty.
Logic error: info returned by check for add-on updates can be None (update check has failed somehow), an empty records list (no updates), or al ist of records with at least one record in it (updates available). At least the first condition was checked but not others, causing NVDA to say that ad-on updates are being downloaded in the background when no updates exist. Therefore make sure that update records list is: not None and not empty.