Add-on GUI: call gui.addonGui.promptUserForRestart instead of duplicating function content.
As part of compatibility info processing work in 2019 (NV Access), add-on GUI module from NVDA Core added 'prompt for restart' function to present restart edialog when add-ons changed, and this add-on wasn't aware of it for over three years. Therefore as part of code cimplification, use the function coming from NVDA Core. Note that it must be called after a delay (100 milliseconds for now) so the host dialog (add-on updates dialog/update progress dialog) can be destroyed first.