Add-on update protocols: define a dedicated method to obtain actual add-on downoad URL from a different source URL.
Seen in all protocols except protocol 4: add-on source may record update URL's not ending in (or not including) '.nvda-addon'. Therefore use a dedicated method (defined in protocol 0) to obtain the actual download link from the source URL. This method resembles getAddonsData except it returns a URL, not add-ons data. This is not needed in protocol 4 as the source contains the actual downlaod URL.