Update protocols: introduce protocol 2.1 (enhanced community-based compatiblity info).
In late August 2022, a change was made to add-on compatiblity JSON to add add-on update information such as version, URL, and hash values for each update channel. This bumps protocol 2 to 2.1 but no separate protocol clas is implemented as the protocol still uses parts of protocol 2.0 (original implementation). In some ways, the enhanced protocol borrows from Spanish community add-ons catalog (protocol 3) and ongoing NV Access datastore work (protocol 4, to be reintroduced later).