Update protocol 2.1: fall back to add-on files repo if compatibility JSON does not come with download URL for a given add-on/channel.
For compatibility, add-on compatibility JSON will use add-on key as channel name for releases/channels. This is useful if Add-on Updater must fall back to obtaining download link from add-on files repository, which is the procedure for when update channels dictionary is not defined in compatibility JSON. Note that not all add-ons have update channels dictionaries, and not all update channels have download links for all released channels as of August 2022.