Add-on dowload/GUI: convert downoad dialog loop to concurrent.futures context.
Use a future to download add-on updates in the background and display progress (although the downloads list will be random). Specifically, a collection of futures will be stored, each future being a key to a dictionary that records destination path and add-on record object (needed for logging purposes). Also, use a try/except statement to catch exceptions coming from future.result() function.