Add-on update record: introduce 'installed channel'. Re #15.
Part of update availability rewrite: use a dedicated 'installed channel' text to denote the installed channel for the add-on which is separate from update channel text used in prior releases. Among other things, the new text is used to compare instaled versus update channels, and if they are different and update versions are different, then whatever version is showing wil be offered to users.