Update protocol 4: attempt to work with version number if defined.
Temporarily store both version name (string) and version number (tuple of major.minor.patch) in addon.version if the latter is present. This will be restored in the reworked 'check for add-on update' method for this protocol. This is a necessary change  since NV Access datastore do record version numbers for add-on releases.