Protocol 4: rework latter part of add-on update check method.
Instead of just returning add-ons with available updates, loop through current add-ons, fetch update info, restore addon.version, then call 'updateAvailable' method with version number argument. If there is no entry for the add-on in the datastore, addon.version remains a string as this is set when update records are first gathered before the protocol is invoked.