Protocols 1 and 2: add-on version parser takes strictly a string argument iwth instlaled add-on version being default.
Add-on version is a string but add-on version parser can return either a string or a None, causing Mypy to report assignment error. Resolve this once and for all by telling the parser to take in strictly a string argument for fallback version, defaulting to installed add-on version. With this change, from add-on info fetcher methods in protocols 1 and 2, there is no need to specify fallback version when parsing add-on versions.