Add-on utilities: introduce cache hash key.
NV Access add-on store introduced 'cache hash' endpoint in June 2023, intended to return the latest Git commit for the store views branch to avoid unnecessary database lookup if the hashes are the same. Therefore use this endpoint in the add-on to avoid json traversal, and in preparation for that, introduce a 'cache hash' key in add-on update state to store store views branch hash. This will be used in update protocols 1 (add-on files get.php/unused), 2 (community add-ons website a.k.a. get.php and compatibility data), and 4 (add-on store).